% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\name{calc_abc2}
\alias{calc_abc2}
\title{2系のABC計算用の関数}
\usage{
calc_abc2(ccdata, BT = 0.8, PL = 0.7, PB = 0, tune.par = c(0.5,
  0.4, 0.4), AAV = NULL, n.catch = 5)
}
\arguments{
\item{ccdata}{year,cpue,catchのデータフレーム}

\item{BT}{initial target level}

\item{PL}{BL = PL*BT}

\item{PB}{BB = PB*BT}

\item{tune.par}{tuning parameters: (beta, delta, lambda)}

\item{AAV}{if this value is given, given value is used}

\item{n.catch}{period for averaging the past catches}
}
\description{
2系のABC計算用の関数
}
\examples{
library(frasyr23)
catch <- c(10,5,10,3,2,1,3)
cpue <- c(11,10,2,3,2,5,2)
example_data <- data.frame(year=1:7,cpue=cpue,catch=catch)

# 2系
example_abc2 <- calc_abc2(example_data)
example_abc2$ABC
# $ABC=1.894191
graph_abc2 <- plot_abc2(example_abc2)

}
