% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\name{calc_abc3}
\alias{calc_abc3}
\title{3系のABC計算用の関数}
\usage{
calc_abc3(ccdata, BT = 0.05, PL = 8, PB = 20, tune.par = c(3.5,
  3.5), n.catch = 5)
}
\arguments{
\item{BT}{initial target level}

\item{PL}{BL = PL*BT}

\item{PB}{BB = PB*BT}

\item{tune.par}{tuning parameters: (beta, delta)}

\item{n.catch}{period for averaging the past catches}

\item{catch}{catch timeseries data}
}
\description{
3系のABC計算用の関数
}
\examples{

library(frasyr23)

catch <- c(15,20,13,14,11,10,5,10,3,2,1,3)
cpue <- c(10,9,8,4,8,11,10,2,3,2,5,2)
example_data <- data.frame(year=2001:2012,cpue=cpue,catch=catch) 

example_abc3 <- calc_abc3(example_data)
graph_abc3 <- plot_abc3(example_abc3)

}
