% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\docType{package}
\name{frasyr23-package}
\alias{frasyr23}
\alias{frasyr23-package}
\title{2,3系のABC計算をするためのパッケージ}
\description{
Calculating ABC for data poor fisheries stocks (type2: only with CPUE and catch, type3: only with catch)
}
\author{
\strong{Maintainer}: Hiroshi Okamura \email{okamura@fra.affrc.go.jp}

}
