% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\name{plot_abc2_fixTerminalCPUE_seqOut}
\alias{plot_abc2_fixTerminalCPUE_seqOut}
\title{2系のABC計算をBTyearオプションありで計算（資源水準導出のためのCPUE時系列制御）した結果をBTyearなしの場合まで連続的に比較してプロットするための関数}
\usage{
plot_abc2_fixTerminalCPUE_seqOut(
  res,
  stock.name = NULL,
  fishseason = 0,
  abc4 = FALSE,
  fillarea = FALSE,
  cpueunit = "",
  catchunit = "（トン）",
  RP = TRUE,
  leftalign = FALSE,
  hcrdist = FALSE
)
}
\arguments{
\item{res}{calc_abc2の返り値、ただしBTyear!=NULL}

\item{fishseason}{X軸のラベルを変更（0なら年、1なら漁期年)}

\item{abc4}{北海道東部の跨り資源で使用する図を描画（TRUEなら使用、デフォルトはFALSE）}

\item{fillarea}{資源量指標値の図にkobeプロットに似た色を塗る（TRUEなら塗る、デフォルトはFALSE）}

\item{cpueunit}{資源量指標値の縦軸見出しに追記したい指標値の単位（例えば"（トン/網）"のように指定する）}

\item{catchunit}{漁獲量トレンドの縦軸見出しに追記したい単位（例えば"（万トン）"のように指定する、デフォルトは"（トン）"）}

\item{RP}{資源量指標値/年のプロットでReference Point（目標・限界管理基準線）を載せる・載せない（デフォルトはTRUE、FALSEでは直近年の資源量指標値をポイントでハイライトする）}

\item{leftalign}{資源量指標値の時系列の長さが漁獲量に比べて短い時、データが無い範囲の空間を削除する（TRUEなら使用、デフォルトはFALSE）}

\item{catchdividedegit}{漁獲量トレンドの縦軸を任意の桁で割る（例えば"catchdividedegit=2"で100で割るように指定する、デフォルトはNULLでそのまま）}
}
\description{
2系のABC計算をBTyearオプションありで計算（資源水準導出のためのCPUE時系列制御）した結果をBTyearなしの場合まで連続的に比較してプロットするための関数
}
