% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\name{plot_abc3}
\alias{plot_abc3}
\title{3系のABC計算結果のをプロットするための関数}
\usage{
plot_abc3(
  res,
  stock.name = NULL,
  fishseason = 0,
  detABC = 0,
  proposal = TRUE,
  catchunit = "(トン)"
)
}
\arguments{
\item{res}{calc_abc3の返り値}
}
\description{
3系のABC計算結果のをプロットするための関数
}
