% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\name{type2_func_empir}
\alias{type2_func_empir}
\title{2系資源計算を経験分布で計算する時のCPUEデータに対してABCを返す関数}
\usage{
type2_func_empir(
  cD,
  cpue,
  simple = FALSE,
  BT = 0.8,
  PL = 0.7,
  PB = 0,
  AAV = 0.4,
  tune.par = c(0.5, 0.5, 0.4),
  beta = 1
)
}
\arguments{
\item{cD}{cpueの値}

\item{cpue}{cpue時系列}

\item{simple}{旧2系的な経験分布にする場合TRUE}

\item{BT}{目標水準}

\item{PL}{目標水準に対する限界水準の割合}

\item{PB}{目標水準に対する禁漁水準の割合}

\item{AAV}{AAV}

\item{tune.par}{チューニングパラメタ}

\item{beta}{保守的漁獲の割合}
}
\description{
2系資源計算を経験分布で計算する時のCPUEデータに対してABCを返す関数
}
