% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_23.r
\name{type2_func_empir_wrapper}
\alias{type2_func_empir_wrapper}
\title{2系を経験分布で計算する時、CPUEの確率点に対して連続的にABCを返す関数}
\usage{
type2_func_empir_wrapper(DL, cpue, simple, type = NULL, ...)
}
\description{
2系を経験分布で計算する時、CPUEの確率点に対して連続的にABCを返す関数
}
